QDataStream about Map and Vector passed test

+++ Server/12308_Server/include/vector.hpp
@@ -492,12 +492,19 @@ template <typename T> class vector {
 };
 }
 
+/*
+ *
+ * Writed by Jingxiao Chen.
+ *
+ *  [Passed test]
+ *
+ */
 #include <QDataStream>
 
 template<class T>
 QDataStream& operator << (QDataStream& out, const ttd::vector<T>&v)
 {
-	out << (quint16)v.size();
+	out << (quint32)(v.size());
 	for (size_t i = 0; i < v.size(); ++i)
 		out << v[i] ;
 	return out;
@@ -507,7 +514,7 @@ template<class T>
 QDataStream& operator >> (QDataStream& in, ttd::vector<T>&v)
 {
 	v.clear();
-	int tmp_n;T d;
+	quint32 tmp_n;T d;
 	in >> tmp_n;
 	v = ttd::vector<T>( tmp_n );
 	for( int i = 0; i < tmp_n; ++i )
